updates database

+++ src/database/migrations/2004_01_01_000020_forms_create_table_form.php
@@ -12,19 +12,21 @@ class FormsCreateTableForm extends Migration {
      */
     public function up()
     {
-        Schema::create('004_401_form', function(Blueprint $table)
+        if(!Schema::hasTable('004_401_form'))
         {
-            $table->engine = 'InnoDB';
-            $table->increments('id_401')->unsigned();
-            $table->string('name_401', 100);
-            $table->integer('email_account_401')->unsigned();
-            $table->boolean('push_notification_401')->default(false);
-            $table->integer('n_unopened_401')->unsigned()->default(0);
+            Schema::create('004_401_form', function (Blueprint $table) {
+                $table->engine = 'InnoDB';
+                $table->increments('id_401')->unsigned();
+                $table->string('name_401', 100);
+                $table->integer('email_account_401')->unsigned();
+                $table->boolean('push_notification_401')->default(false);
+                $table->integer('n_unopened_401')->unsigned()->default(0);
 
-            // correos de reenvío
-            $table->foreign('email_account_401', 'fk01_004_401_form')->references('id_013')->on('001_013_email_account')
-                ->onDelete('restrict')->onUpdate('cascade');
-        });
+                // correos de reenvío
+                $table->foreign('email_account_401', 'fk01_004_401_form')->references('id_013')->on('001_013_email_account')
+                    ->onDelete('restrict')->onUpdate('cascade');
+            });
+        }
     }
 
     /**
