+++ src/database/migrations/2004_01_01_000050_forms_create_table_comment.php
@@ -12,21 +12,23 @@ class FormsCreateTableComment extends Migration {
      */
     public function up()
     {
-        Schema::create('004_404_comment', function(Blueprint $table)
+        if(!Schema::hasTable('004_404_comment'))
         {
-            $table->engine = 'InnoDB';
-            $table->increments('id_404')->unsigned();
-            $table->integer('record_404')->unsigned();
-            $table->integer('user_404')->unsigned();
-            $table->integer('date_404')->unsigned();
-            $table->string('subject_404', 255);
-            $table->text('comment_404');
+            Schema::create('004_404_comment', function (Blueprint $table) {
+                $table->engine = 'InnoDB';
+                $table->increments('id_404')->unsigned();
+                $table->integer('record_404')->unsigned();
+                $table->integer('user_404')->unsigned();
+                $table->integer('date_404')->unsigned();
+                $table->string('subject_404', 255);
+                $table->text('comment_404');
 
-            $table->foreign('record_404', 'fk01_004_404_comment')->references('id_403')->on('004_403_record')
-                ->onDelete('cascade')->onUpdate('cascade');
-            $table->foreign('user_404', 'fk02_004_404_comment')->references('id_010')->on('001_010_user')
-                ->onDelete('restrict')->onUpdate('cascade');
-        });
+                $table->foreign('record_404', 'fk01_004_404_comment')->references('id_403')->on('004_403_record')
+                    ->onDelete('cascade')->onUpdate('cascade');
+                $table->foreign('user_404', 'fk02_004_404_comment')->references('id_010')->on('001_010_user')
+                    ->onDelete('restrict')->onUpdate('cascade');
+            });
+        }
     }
 
     /**
